worker/requirements.txt
beautifulsoup4==4.12.2
celery==5.2.7
httpx==0.24.1
lxml==4.9.2
pandas==2.0.2
python-dotenv==1.0.0
redis==4.5.5
selenium==4.9.1
